separate tests for experimental features
