+++ test/data/test_utils.py
@@ -1,7 +1,6 @@
 import io
 from torchtext.data import get_tokenizer
 from torchtext.utils import unicode_csv_reader
-from torchtext.experimental.functional import ngrams_func
 from ..common.torchtext_test_case import TorchtextTestCase
 from ..common.assets import get_asset_path
 
@@ -49,17 +48,3 @@ class TestUtils(TorchtextTestCase):
                 ref_lines.append(line)
 
         self.assertEqual(ref_lines, test_lines)
-
-    def test_ngrams_func(self):
-        func = ngrams_func(1)
-        assert func(['A', 'string', 'particularly', 'one', 'with', 'slightly']) == \
-            ['A', 'string', 'particularly', 'one', 'with', 'slightly']
-        func = ngrams_func(2)
-        assert func(['A', 'string', 'particularly', 'one', 'with', 'slightly']) == \
-            ['A', 'string', 'particularly', 'one', 'with', 'slightly', 'A string', 'string particularly',
-             'particularly one', 'one with', 'with slightly']
-        func = ngrams_func(3)
-        assert func(['A', 'string', 'particularly', 'one', 'with', 'slightly']) == \
-            ['A', 'string', 'particularly', 'one', 'with', 'slightly', 'A string', 'string particularly',
-             'particularly one', 'one with', 'with slightly', 'A string particularly',
-             'string particularly one', 'particularly one with', 'one with slightly']
